+++ app/components/Views/confirmations/components/info/musd-max-conversion-info/musd-max-conversion-asset-header/index.ts
@@ -1,0 +1,5 @@
+export {
+  MusdMaxConversionAssetHeader,
+  MusdMaxConversionAssetHeaderSkeleton,
+  MusdMaxConversionAssetHeaderTestIds,
+} from './musd-max-conversion-asset-header';
