+++ app/components/Views/confirmations/components/info/musd-max-conversion-info/musd-max-conversion-asset-header/musd-max-conversion-asset-header.styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { Theme } from '../../../../../../util/theme/models';
+import { Theme } from '../../../../../../../util/theme/models';
 
 const styleSheet = (params: { theme: Theme }) => {
   const { theme } = params;
